Removed django.contrib.gis
